contrib/wordlyworks/requirements.txt
loguru==0.7.0
python-dotenv==1.0.0
TheNounProjectAPI==1.0.5
